feat(api): change endpoints to reflect CX-124
This PR also splits the existing types into their own respective file. Note that this PR uses the new endpoints, see CX-124 for the description.
