+++ src/types/party.ts
@@ -1,0 +1,11 @@
+/*
+  This program has been developed by students from the bachelor Computer Science at Utrecht
+  University within the Software Models course.
+
+  © Copyright Utrecht University (Department of Information and Computing Sciences)
+*/
+
+export type Party = {
+    id: string,
+    name: string
+}
