+++ src/types/project.ts
@@ -4,30 +4,18 @@
 
   © Copyright Utrecht University (Department of Information and Computing Sciences)
 */
+import {Person} from "@/types/person.ts";
+import {Product} from "@/types/product.ts";
+import {Party} from "@/types/party.ts";
+
 
 export type Project = {
     id: string,
     title: string,
     description: string,
-    startDate: Date,
-    endDate: Date,
+    start_date: Date,
+    end_date: Date,
     people: Person[],
     products: Product[],
     parties: Party[]
-}
-
-export type Person = {
-    id: string,
-    name: string
-}
-
-export type Party = {
-    id: string,
-    name: string
-}
-
-export type Product = {
-    id: string,
-    title: string,
-    url: string
 }
